% Generated by roxygen2: do not edit by hand
% Please edit documentation in others/array2list.R
\name{array2list}
\alias{array2list}
\title{Convert an array (or a matrix) to a list}
\usage{
array2list(X, MARGIN)
}
\arguments{
\item{X}{"array"}

\item{MARGIN}{"integer"}
}
\value{
"list"
}
\description{
Also works with matrices.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
