% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attached.packages.R
\name{attached.packages}
\alias{attached.packages}
\title{list all attached packages}
\usage{
attached.packages()
}
\value{
NA
}
\description{
List all attached packages
}
\note{
Created: Wed Mar 21 12:01:21 CET 2012;
      Current: Wed Mar 21 12:01:27 CET 2012.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
