% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist__beta2.R
\name{beta2}
\alias{beta2}
\alias{rbeta2}
\alias{qbeta2}
\alias{dbeta2}
\title{The reparameterized beta distribution}
\usage{
rbeta2(n, mean, sd, ncp = 0, cond.warning = TRUE)

qbeta2(
  p,
  mean,
  sd,
  ncp = 0,
  lower.tail = TRUE,
  log.p = FALSE,
  cond.warning = TRUE
)

dbeta2(x, mean, sd, ncp = 0, log = FALSE, cond.warning = TRUE)
}
\arguments{
\item{n}{Number of random draws.}

\item{mean}{"numeric" The mean value of beta distribution.}

\item{sd}{"numeric" The standard deviation of beta distribution.}

\item{ncp}{Non-centrality parameter passed to beta distribution functions.}

\item{cond.warning}{"logical" If warning should be printed if NA produced.}

\item{lower.tail, log.p}{Logical flags passed to beta distribution functions.}

\item{x, p}{Quantiles or probabilities.}

\item{log}{Logical; if `TRUE`, return log density.}
}
\value{
See the return of beta distribution.
}
\description{
The reparameterized beta distribution. This is a wrapper of the usual beta functions.
}
\details{
See the help for general information of beta distribution. Note that the
mean and standard deviation should be in (0, 1) and (0, mean(1-mean)),
respectively.
}
\note{
Created: Tue Apr 10 19:15:06 CEST 2012;
      Current: Tue Apr 10 19:15:11 CEST 2012.
}
\references{
Li,  2012
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
