% Generated by roxygen2: do not edit by hand
% Please edit documentation in others/block.diag.R
\name{block.diag}
\alias{block.diag}
\title{Block diagonal matrix.}
\usage{
block.diag(x)
}
\arguments{
\item{x}{"list" contains the block matrices}
}
\value{
"matrix" The block diagonal matrix.
}
\description{
Create a block diagonal matrix.
}
\details{
For very large matrix. Consider using .bdiag() function in package "Matrix".
}
\note{
Created: Thu Feb 03 19:16:21 CET 2011;
      Current: Sun Mar 04 17:14:51 CET 2012.
}
\references{
Modified from R-help list.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
