% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.partition.R
\name{data.partition}
\alias{data.partition}
\title{Partition data index for cross validation.}
\usage{
data.partition(nObs, args)
}
\arguments{
\item{nObs}{"positive numeric"}

\item{args}{list N: no. of subsets,  method: how to partition}
}
\value{
"list" The prediction subsets.
}
\description{
Partition data index for cross validation.
}
\note{
First version: Mon Sep 20 21:08:01 CEST 2010;
      Current:       Mon Sep 20 21:08:11 CEST 2010.
}
\references{
Li Villani Kohn 2010
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
