% Generated by roxygen2: do not edit by hand
% Please edit documentation in dist/DensGradHess.R
\name{DensGradHess}
\alias{DensGradHess}
\title{A collection of gradient and Hessian for the log form of common densities.}
\usage{
DensGradHess(B, ..., type = "norm", grad = TRUE, Hess = TRUE)
}
\arguments{
\item{B}{"matrix".  The paramter that need to be added with a prior. The gradient and
hessian are calculated conditional on B. B should be always an one-column matrix}

\item{...}{The arguments for the densities. The name should be matched exactly.}

\item{type}{density type}

\item{grad}{"logical" Should the gradient be computed?}

\item{Hess}{"logical" Should the Hessian be computed?}
}
\value{
"list". The gradient and Hessian (if required) matrices, see below. grad is an
    one-column matrix; Hess is a squared matrix whose dimension on row and column are
    same as length of B
}
\description{
The parameters after "..." should be matched exactly.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
