% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others__diagAlg.R
\name{diag1}
\alias{diag1}
\title{Fast identity matrix}
\usage{
diag1(n)
}
\arguments{
\item{n}{A positive integer giving the matrix dimension.}
}
\value{
An `n` by `n` identity matrix.
}
\description{
Fast identity matrix
}
