% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart.R
\name{diwishart}
\alias{diwishart}
\title{Density for the inverse wishart distribution}
\usage{
diwishart(X, df, V, log = TRUE)
}
\arguments{
\item{X}{NA}

\item{df}{NA}

\item{V}{Location matrix}

\item{log}{TRUE or FALSE}
}
\description{
Details.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
