% Generated by roxygen2: do not edit by hand
% Please edit documentation in dgp/rmixmorm.R
\name{dmixnorm}
\alias{dmixnorm}
\title{Density for normal mixtures}
\usage{
dmixnorm(x, means, sigmas, weights, log = FALSE)
}
\arguments{
\item{means}{A `q-by-k` matrix. Mean value vector within each component in total k
components.}

\item{sigmas}{A `q-by-q-by-k` array.  Variance covariance matrix with in each
component.}

\item{weights}{A `k`-length vector.  The weight in each component.}

\item{n}{An positive integer. Numbers of samples to be generated.}
}
\value{
An `n-by-q` matrix
}
\description{
Density for univariate and multivariate mixture normal distribution
}
\details{
Density from mixture of normals.
}
\examples{
n <- 1
means <- matrix(c(-5, 0, 5), 1)
sigmas <- array(c(1, 1, 1), c(1, 1, 3))
weights <- c(0.3, 0.4, 0.3)
out <- rmixnorm(n, means, sigmas, weights)
out.density <- dmixnorm(out, means, sigmas, weights)
}
\references{
Villani et al 2009
}
\author{
Feng Li, Central University of Finance and Economics.
}
