% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmixmorm.R
\name{dmixnorm.ts}
\alias{dmixnorm.ts}
\title{The conditional density for AR-type mixture of normals}
\usage{
dmixnorm.ts(y, means.ar.par.list, sigmas.list, weights, log = FALSE)
}
\arguments{
\item{y}{An `n`-length vector. The time series.}

\item{means.ar.par.list}{A `k`-length list. Each element in the list consists the
coefficients in the AR process.}

\item{sigmas.list}{A `k`-length lilst.}

\item{weights}{A `k`-length vector. The weight in each component.}

\item{log}{Logical; If TRUE, the log density is returned.}
}
\description{
For more information see dmixnorm()
}
\examples{
n = 1000
means.ar.par.list = list(c(0, 0.8), c(0, 0.6, 0.3))
require("fGarch")
sigmas.spec <- list(garchSpec(model = list(alpha = c(0.05, 0.06)), cond.dist = "norm"),
                    garchSpec(model = list(alpha = c(0.05, 0.05)), cond.dist = "norm"))
sigmas.list <- lapply(lapply(sigmas.spec, garchSim, extended = TRUE, n = n),
                      function(x) x$sigma)
weights <- c(0.8, 0.2)
y = rmixnorm.ts(n = n, means.ar.par.list = means.ar.par.list, sigmas.list = sigmas.list,
                weights = weights)
out = dmixnorm.ts(y = y, means.ar.par.list = means.ar.par.list,
                  sigmas.list = sigmas.list, weights = weights, log = TRUE)
}
