% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbg__dot_ls.objects.R
\name{.ls.objects}
\alias{.ls.objects}
\title{List the object size
Modified from Dirk Eddelbuettel's version
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
improved list of objects}
\usage{
.ls.objects(
  pos = 1,
  pattern,
  order.by,
  decreasing = FALSE,
  head = FALSE,
  n = 5
)
}
\arguments{
\item{pos}{Environment position passed to `ls()` and `get()`.}

\item{pattern}{Optional object name pattern passed to `ls()`.}

\item{order.by}{Optional output column name used for sorting.}

\item{decreasing}{Logical; if `TRUE`, sort decreasing.}

\item{head}{Logical; if `TRUE`, return only the first `n` rows.}

\item{n}{Number of rows returned when `head = TRUE`.}
}
\value{
A data frame with object type, size, rows, and columns.
}
\description{
List the object size
Modified from Dirk Eddelbuettel's version
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
improved list of objects
}
