% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitt.R
\name{dsplitt}
\alias{dsplitt}
\title{Split student-t distribution}
\usage{
dsplitt(x, mu, df, phi, lmd, log)
}
\arguments{
\item{x}{"numeric".}

\item{mu}{"numeric".

       Location parameter. The mode of the density}

\item{df}{"positive number".

       Degrees of freedom.}

\item{phi}{"positive number".

       Scale parameter.}

\item{lmd}{"positive number".

       Skewness parameter. If is 1, reduced to symmetric student t
       distribution.}

\item{log}{"logical".

       If the vale is TRUE, return the logarithm from.}
}
\value{
"numeric"

        Return the density of split-t distribution.
}
\description{
The split-t distribution.
}
\note{
Created: Sat Jan 05 22:49:34 CET 2013;

      Current: Sat Jan 05 22:49:39 CET 2013.
}
\references{
Li,  Villani and  Kohn (2010)
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
