% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flutils-package.R
\docType{package}
\name{flutils-package}
\alias{flutils}
\alias{flutils-package}
\title{flutils: Utility Functions for Statistical Modeling}
\description{
A collection of utility functions for simulation, distributions, matrix
algebra, splines, MCMC helpers, and package development.
}
\keyword{internal}
