% Generated by roxygen2: do not edit by hand
% Please edit documentation in others/foldMat.R
\name{foldMat}
\alias{foldMat}
\title{Fold a matrix}
\usage{
foldMat(Mat, nfolds, byrow = TRUE)
}
\arguments{
\item{Mat}{matrix}

\item{nfolds}{number of folds}

\item{byrow}{fold by row or by column.}
}
\value{
new matrix
}
\description{
This function fold a matrix into different shape
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
