% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghypergeo.R
\name{ghypergeo}
\alias{ghypergeo}
\title{The generalized hypergeometric function pFq.}
\usage{
ghypergeo(a, b, z, k = 10)
}
\arguments{
\item{a}{"n-by-p matrix"}

\item{b}{"n-by-q matrix"}

\item{z}{"positive vector of length one or n"}

\item{k}{"positive integer"

       Indicates how long of the series should be used. The default value
       is 10.}
}
\value{
"n-by-1 matrix"

      The matrix of length n.
}
\description{
The results are evaluated via the series expansion of Pochhammer symbols.
}
\note{
Created: Tue Jan 15 12:36:41 CET 2013;
      Current: Sat May 30 15:07:07 CST 2015.
}
\references{
HypergeometricPFQ in Mathematica 10
}
\author{
Feng Li, Central University of Finance and Economics.
}
