% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others__diagAlg.R
\name{\%d*\%}
\alias{\%d*\%}
\alias{\%*d\%}
\alias{\%d*d\%}
\alias{tMdN}
\title{Fast matrix algebra with diagonal matrices.}
\usage{
d \%d*\% M

M \%*d\% d

d \%d*d\% M

tMdN(M, d, N = M)
}
\arguments{
\item{d}{"vector" Diagonal elements form a diagonal matrix.}

\item{M}{"matrix" Dense matrix that is conformable to matrix diag(d).}

\item{N}{A dense matrix used by `tMdN()`.}
}
\value{
New conformable matrix.
}
\description{
This implementation does not involve sparse matrix structure.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
