% Generated by roxygen2: do not edit by hand
% Please edit documentation in plots/grid2.R
\name{grid2}
\alias{grid2}
\title{Add grids on plot}
\usage{
grid2(x.at = NA, y.at = NA, col = "black", lty = "dotted", lwd = 0.5, ...)
}
\arguments{
\item{x.at}{position on x-axis}

\item{y.at}{position on y-axis}

\item{col}{color}

\item{lty}{line type}

\item{lwd}{line width}

\item{...}{other arguments passed to abline()}
}
\value{
NA
}
\description{
Better than grid() function with abline() implementation
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
