% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math__I.Kron.R
\name{I.Kron}
\alias{I.Kron}
\title{Block diagonal Kronecker product with an identity matrix}
\usage{
I.Kron(n, X)
}
\arguments{
\item{n}{A positive integer.}

\item{X}{A matrix.}
}
\value{
A block diagonal matrix with `n` copies of `X`.
}
\description{
Construct `kronecker(diag(n), X)`.
}
