% Generated by roxygen2: do not edit by hand
% Please edit documentation in math/ibeta.R
\name{ibeta}
\alias{ibeta}
\title{Incomplete beta function}
\usage{
ibeta(x, a, b, log = FALSE, reg = FALSE)
}
\arguments{
\item{x}{a numeric vector in the range [0,1], the point at which the incomplete Beta
function is evaluated.}

\item{a}{non-negative numeric vectors, the first parameter of the incomplete beta function.}

\item{b}{non-negative numeric vectors, the second parameter of the incomplete beta function.}

\item{log}{logical; If TRUE, return the log incomplete beta function.}

\item{reg}{logical; If TRUE, return the regularized incomplete beta function. Same as
the Mathematica definition.}
}
\value{
a numeric vector.
}
\description{
This function is used to compute the incomplete beta function.
}
\references{
\url{http://mathworld.wolfram.com/IncompleteBetaFunction.html}

TODO: let -1 < b < oO.
}
