% Generated by roxygen2: do not edit by hand
% Please edit documentation in mcmc/ineff.R
\name{ineff}
\alias{ineff}
\title{Inefficiency factor}
\usage{
ineff(par)
}
\arguments{
\item{par}{"vector" like}
}
\value{
"numeric"
}
\description{
Inefficiency factor of given MCMC chain
}
\details{
Just the sum of autocorrelations
}
\note{
Created: Sun Sep 02 13:59:26 CEST 2012; Current: Tue Aug 11 16:17:18 CST 2015.
}
\references{
Geweke 1992
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
