% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.displayopen.R
\name{is.displayopen}
\alias{is.displayopen}
\title{is.displayopen}
\usage{
is.displayopen()
}
\value{
TRUE/FALSE
}
\description{
Is display open or not.
}
\details{
A simple way to check if the display is available.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
