% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others__is.positivedefinite.R
\name{is.positivedefinite}
\alias{is.positivedefinite}
\title{Test positive definiteness}
\usage{
is.positivedefinite(X)
}
\arguments{
\item{X}{A square numeric matrix.}
}
\value{
`TRUE` if all eigenvalues are strictly positive; otherwise `FALSE`.
}
\description{
Test positive definiteness
}
