% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others__is.singular.R
\name{is.singular}
\alias{is.singular}
\title{Check if a square matrix is computational singular}
\usage{
is.singular(X, tol = .Machine$double.eps * 1000)
}
\arguments{
\item{X}{A square matrix or Matrix object.}

\item{tol}{Numeric tolerance below which the reciprocal condition number is
treated as singular.}
}
\value{
TRUE/FALSE
}
\description{
Check if a square matrix is computational singular
}
\note{
Created: Mon Jan 02 17:58:12 CET 2012;
      Current: Mon Jan 02 17:58:37 CET 2012.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
