% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math__K.R
\name{K}
\alias{K}
\title{Commutation matrix}
\usage{
K(m, n)
}
\arguments{
\item{m, n}{Positive integer matrix dimensions.}
}
\value{
A commutation matrix with dimension `m * n` by `m * n`.
}
\description{
Construct the `mn` by `mn` commutation matrix.
}
