% Generated by roxygen2: do not edit by hand
% Please edit documentation in mcmc/make_stepsize.R
\name{make_stepsize}
\alias{make_stepsize}
\title{Make step size for SGLD}
\usage{
make_stepsize(steprange, n, args)
}
\arguments{
\item{steprange}{Range for steps.}

\item{n}{Total steps.}

\item{args}{Other arguments}
}
\value{
Steps
}
\description{
Step size is also known as the learning rate in stochastic gradient descent
algorithms.
}
\author{
Feng Li, Central University of Finance and Economics.
}
