% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math__colSds.R
\name{matrix-summaries}
\alias{matrix-summaries}
\alias{colSds}
\alias{rowSds}
\alias{colVars}
\alias{rowVars}
\title{Row and column summaries}
\usage{
colSds(x)

rowSds(x)

colVars(x)

rowVars(x)
}
\arguments{
\item{x}{A numeric matrix.}
}
\value{
A numeric vector.
}
\description{
Convenience wrappers around `apply()` for row and column standard deviations
and variances.
}
