% Generated by roxygen2: do not edit by hand
% Please edit documentation in others/mesh.grid.R
\name{mesh.grid}
\alias{mesh.grid}
\title{Same as expand.grid function but return a matrix.}
\usage{
mesh.grid(x1, x2 = x1)
}
\arguments{
\item{x1}{first grid}

\item{x2}{second grid}
}
\value{
matrix
}
\description{
Simple implementation to expand grid
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
