% Generated by roxygen2: do not edit by hand
% Please edit documentation in others/mseq.R
\name{mseq}
\alias{mseq}
\title{A simple sequence generator}
\usage{
mseq(init.seq, init.length, length.out)
}
\arguments{
\item{init.seq}{"vecter" the initial sequence without zero}

\item{init.length}{"vecter" initial circle length,  including zeros}

\item{length.out}{"integer" how many circles to be generated}
}
\value{
"vecter"
}
\description{
mainly used in the gradient part
}
\note{
First version: ; Current: .
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
