% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvgamma.R
\name{mvgamma}
\alias{mvgamma}
\title{Multivariate gamma function}
\usage{
mvgamma(p, x, log = TRUE)
}
\arguments{
\item{p}{"integer".  The dimension of the multivariate gamma function.}

\item{x}{"vector"
A vector or sclor can be used. See also gamma() function.}

\item{log}{"logical"
If TRUE, logarithm of multivariate gamma function will be returned.}
}
\value{
Same length as "x" has.
}
\description{
The multivariate gamma function is computed by using it's recursive
definition or product form. Use apply() on top of the mvgamma function if p
is also a vector.
}
\note{
Initial: Mon Sep 20 23:54:20 CEST 2010;
      Current: Tue Jan 15 18:56:48 CET 2013.

TODO: Allow p be a vector of same length of x
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
