% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVStocks.R
\name{MVStocks}
\alias{MVStocks}
\title{Multivariate Variables for Stocks}
\usage{
MVStocks(
  from,
  to,
  stocks = c("^SML", "^OEX"),
  StdDataMethod,
  save2diskPath,
  ...
)
}
\arguments{
\item{from}{integer format as YYYYMMDD}

\item{to}{integer format as YYYYMMDD}

\item{stocks}{a vector of characters. The aberrations are same as stocks used in
Yahoo! Finance.}

\item{StdDataMethod}{character. Whether the data should be standardized.}

\item{save2diskPath}{Is not missing, the result will be saved to the path provided
here.}

\item{...}{Other arguments passed to stock2covarites}
}
\value{
list or save to disk.
}
\description{
Download and build multivariate responses and covariates from Yahoo Finance.
}
\references{
Li Villani 2012
}
\author{
Feng Li, Central University of Finance and Economics.
}
