% Generated by roxygen2: do not edit by hand
% Please edit documentation in mcmc/parLinkFun.R
\name{parLinkFun}
\alias{parLinkFun}
\title{Transform the mean function in the GLM framework.}
\usage{
parLinkFun(mu, linkArgs)
}
\arguments{
\item{mu}{mean}

\item{linkArgs}{linking arguments}
}
\value{
"one-col-matrix" of the same dimension as the linear predictor
}
\description{
This function create the linkage.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
