% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parMeanFun.R
\name{parMeanFun}
\alias{parMeanFun}
\title{Transform the mean function in the GLM framework.}
\usage{
parMeanFun(X, beta, linkArgs)
}
\arguments{
\item{X}{"matrix"}

\item{beta}{"one-col-matrix"}

\item{linkArgs}{linkage arguments}

\item{link}{"character" Type of link function}
}
\value{
"one-col-matrix" of the same dimension as the linear predictor
}
\description{
GLM mean function
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
