% Generated by roxygen2: do not edit by hand
% Please edit documentation in mcmc/parMeanFunGrad.R
\name{parMeanFunGrad}
\alias{parMeanFunGrad}
\title{The gradient for the mean function in the GLM framework.}
\usage{
parMeanFunGrad(par, linkArgs)
}
\arguments{
\item{par}{"vector"}

\item{linkArgs}{"list"}
}
\value{
"matrix" of the same dimension as the linear predictor
}
\description{
<details>
}
\note{
Created: Thu Nov 24 11:46:32 CET 2011;
      Current: Thu Nov 24 11:46:39 CET 2011.
}
\references{
Li 2012
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
