% Generated by roxygen2: do not edit by hand
% Please edit documentation in math/pochhammer.R
\name{pochhammer}
\alias{pochhammer}
\title{Pochhammer symbol}
\usage{
pochhammer(a, n, log)
}
\arguments{
\item{a}{"vector".}

\item{n}{"positive integer vector".}

\item{log}{"logical"

       If TRUE, return the logarithm of Pochhammer symbol. Else return the
usual one.}
}
\value{
"matrix"

       The (i, j)th element of the matrix indicates the Pochhammer symbol
for (a[i])n[j].
}
\description{
The Pochhammer symbol is evaluated via the expression (a)n  =
gamma(a+n)/gamma(a) for positive integer n.
}
\note{
Created: Tue Jan 15 11:20:51 CET 2013;
      Current: Tue Jan 15 11:21:00 CET 2013.
}
\references{
NA
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
