% Generated by roxygen2: do not edit by hand
% Please edit documentation in mcmc/progressbar.R
\name{progressbar}
\alias{progressbar}
\title{Very Simple progress bar for "for loops"}
\usage{
progressbar(iIter, nIter)
}
\arguments{
\item{iIter}{Current iteration}

\item{nIter}{Total iteration}
}
\description{
Call this function within for loop. Not work with while loop
Have not test with windows Gui
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
