% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdist.R
\name{rdist}
\alias{rdist}
\title{Euclidean distance matrix between two matrices}
\usage{
rdist(X, Y, log = TRUE)
}
\arguments{
\item{X}{an `m-by-p` numeric matrix.}

\item{Y}{an `n-by-p` numeric matrix.}

\item{log}{logical; If TRUE, the log distance matrix is returned.}
}
\value{
an `m-by-n` numeric matrix.
}
\description{
Calculate the distance between all rows of X and Y.
}
