% Generated by roxygen2: do not edit by hand
% Please edit documentation in others/rhex.R
\name{rhex}
\alias{rhex}
\title{Random hex}
\usage{
rhex(n)
}
\arguments{
\item{n}{"positive-integer"}
}
\value{
"string"
}
\description{
A simple hexadecimal random string generator
}
\note{
Created: Wed Oct 17 20:04:58 CEST 2012;
      Current: Wed Oct 17 20:05:05 CEST 2012.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
