% Generated by roxygen2: do not edit by hand
% Please edit documentation in dist/wishart.R
\name{riwishart}
\alias{riwishart}
\title{Random draws from Wishart (nu,V) and IW}
\usage{
riwishart(df, V)
}
\arguments{
\item{df}{degrees of freedom}

\item{V}{Parameters.}
}
\value{
NA
}
\description{
Details.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
