% Generated by roxygen2: do not edit by hand
% Please edit documentation in dist/wishart.R
\name{rwishart}
\alias{rwishart}
\title{Random draws from Wishart (nu,V) and IW}
\usage{
rwishart(nu, V)
}
\arguments{
\item{nu}{parameter of Wishart distribution.}

\item{V}{parameter of Wishart distribution.}
}
\value{
NA
}
\description{
Came from package "bayesm", by Peter Rossi
}
