% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc__set.crossvalid.R
\name{set.crossvalid}
\alias{set.crossvalid}
\title{Setup cross-validation}
\usage{
set.crossvalid(nObs, crossValidArgs)
}
\arguments{
\item{nObs}{Number of observations.}

\item{crossValidArgs}{Cross-validation arguments.}
}
\value{
list
}
\description{
Tools to setup cross-validations
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
