% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\title{Print object.size}
\usage{
size(x, unit = "auto")
}
\arguments{
\item{x}{object name}

\item{unit}{the unit to use. Auto default.}
}
\value{
size
}
\description{
Object size in a human-readable way.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
TODO: allow multiple inputs
}
