% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StdData.R
\name{StdData}
\alias{StdData}
\title{StdData}
\usage{
StdData(X, method)
}
\arguments{
\item{X}{"Numercial". The data to be standardized. "X" should be a numeric matrix or
a vector. The standardization methods do _NOT_ apply to tho following situations: 1)
there is only one row in the matrix. 2) some colums have constant values.}

\item{method}{"Numerical". If "stdnorm", X will be standardized with mean zero and
standard deviation 1. If "-1to1", X will be restricted to [-1, 1].  The "method"
applies to each colums for a matrix or a vector.}
}
\value{
A matrix or a vector depends on the input X
}
\description{
Standardize a vector or colums of a matrix
}
\note{
First: Thu Mar 4 13:50:30 CET 2010 Current: Sun Dec 21 04:24:52 EST 2014
}
\author{
Feng Li, Dept. of Statistics, Stockholm University, Sweden.
}
