% Generated by roxygen2: do not edit by hand
% Please edit documentation in dmp/stock2covariates.R
\name{stock2covariates}
\alias{stock2covariates}
\title{Construct time series covariates.}
\usage{
stock2covariates(file, g = c(0.95, 0.8), kappa = 1, ma = c(1, 5, 20))
}
\arguments{
\item{file}{"character".  The location of the CSV format file from Yahoo!  Finance
which can be the http address.}

\item{g}{"vector".  Parameters for geometrically averaging.}

\item{kappa}{"scaler".}

\item{ma}{"vector" with integer entries.  Moving average indicator for the returns .}
}
\value{
"data.frame".  The returned data frame only contains the non-NA data due to
the moving average procedure.
}
\description{
Construct sensible covariates for stock market data.
}
\details{
The historical finance data can be obtained from "Yahoo! Finance".
}
\note{
Created: Thu Jan 05 13:23:35 CET 2012; Current: Mon Sep 28 22:01:20 CST 2015.
}
\references{
Geweke, and Keane (2007) p.274
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
