% Generated by roxygen2: do not edit by hand
% Please edit documentation in math/tr.R
\name{tr}
\alias{tr}
\title{trace of a matrix}
\usage{
tr(X)
}
\arguments{
\item{X}{squire matrix.}
}
\value{
scalor
}
\description{
Give the trace of a squire matrix.
}
\details{
It is essentially "sum(diag(X))" but faster.
}
\note{
First version: Wed Dec 29 16:25:04 CET 2010;
      Current:       Wed Dec 29 16:25:10 CET 2010.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
