% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vech.R
\name{vech}
\alias{vech}
\title{Make a vech vector from a full matrix}
\usage{
vech(X, diag = TRUE)
}
\arguments{
\item{X}{Full matrix}

\item{diag}{TRUE or FALSE}
}
\value{
vector
}
\description{
This function create a vech vector
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
