% Generated by roxygen2: do not edit by hand
% Please edit documentation in math/vech2m.R
\name{vech2m}
\alias{vech2m}
\title{Convert from the vech vector to the full matrix.}
\usage{
vech2m(vech, diag = TRUE)
}
\arguments{
\item{vech}{Vech vector}

\item{diag}{TRUE or FALSE}
}
\value{
Full matrix
}
\description{
Vech to matrix
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
