% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systools__warningsClear.R
\name{warningsClear}
\alias{warningsClear}
\title{Clear recorded warnings}
\usage{
warningsClear(envir = baseenv())
}
\arguments{
\item{envir}{Environment from which `last.warning` should be removed.}
}
\value{
`NULL`, invisibly.
}
\description{
Clear recorded warnings
}
