% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichInd.R
\name{whichInd}
\alias{whichInd}
\title{Raw indices finder for array.}
\usage{
whichInd(arr.ind, dims)
}
\arguments{
\item{arr.ind}{"matrix or vector"
       When it is a matrix, the length of columns should be the as the
length of the "dims". When it is a vector, the argument length should be
the same as the length of "dims".}

\item{dims}{"vector"
       The dimension of the array. Note that the last dimension is not
needed in the calculations,  so if you may just write "NA" there.}
}
\value{
"vector" The raw indices.
}
\description{
Find the raw indices for given indices in array form.
}
\details{
This is exactly the inverse function of `arrayInd()'.
}
\note{
Created: Wed Mar 14 18:03:01 CET 2012;
      Current: Wed Mar 14 18:03:09 CET 2012.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
