% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math__zeta.R
\name{zeta}
\alias{zeta}
\title{Approximate Riemann zeta function}
\usage{
zeta(s, k = 10)
}
\arguments{
\item{s}{Numeric vector of powers.}

\item{k}{Positive integer number of terms in the finite sum.}
}
\value{
Numeric vector with the same length as `s`.
}
\description{
Evaluate a finite-series approximation to the Riemann zeta function.
}
