% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_ls.objects.R
\name{.ls.objects}
\alias{.ls.objects}
\title{List the object size
Modified from Dirk Eddelbuettel's version
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
improved list of objects
TODO: Better default sorting}
\usage{
.ls.objects(
  pos = 1,
  pattern,
  order.by,
  decreasing = FALSE,
  head = FALSE,
  n = 5
)
}
\description{
List the object size
Modified from Dirk Eddelbuettel's version
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
improved list of objects
TODO: Better default sorting
}
