% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagAlg.R
\name{\%d*\%}
\alias{\%d*\%}
\title{Fast matrix algebra with diagonal matrices.}
\usage{
d \%d*\% M
}
\arguments{
\item{d}{"vector" Diagonal elements form a diagonal matrix.}

\item{M}{"matrix" Dense matrix that is conformable to matrix diag(d).}
}
\value{
New conformable matrix.
}
\description{
This implementation does not involve sparse matrix structure.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
