% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.all.R
\name{save.all}
\alias{save.all}
\title{Save all the objects to a folder.}
\usage{
save.all(save.output, ModelDescription)
}
\arguments{
\item{save.output}{If the output should save.}

\item{ModelDescription}{String}

\item{Starting.time}{Computing initialized time}
}
\description{
Save all the objects to a folder.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
